+++ mail/mail-common/src/models/message.rs
@@ -1281,7 +1281,7 @@ impl Message {
                 // address id are updated.
                 // If no record exists we create a new one.
                 let mut result = Vec::with_capacity(self.attachments_metadata.len());
-                for metadata in &self.attachments_metadata {
+                for metadata in &mut self.attachments_metadata {
                     let mut attachment = Attachment::find_first(
                         "WHERE remote_id = ?",
                         params![metadata.remote_id.clone()],
@@ -1295,8 +1295,8 @@ impl Message {
                     attachment.local_message_id = self.local_id;
                     attachment.remote_message_id = self.remote_id.clone();
                     attachment.save(bond).await?;
-
                     let local_id = attachment.local_id.expect("Should be set");
+                    metadata.local_id = Some(local_id);
 
                     bond.execute(
                         "INSERT OR IGNORE INTO message_attachments VALUES (?,?)",