+++ mail/mail-uniffi/CHANGELOG.md
@@ -14,9 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `draft_send_result_mark_read`
   - `draft_send_result_delete`
   - `Draft::send_result` - Loads associated send result with `open_draft` if any is available.
+  - [ET-1192] `ContactEmailItem` has two new fields: `is_proton` & `last_used_time`
 
 ### Changed
+
   - `open_draft` now returns `OpenDraft` type which includes whether the body is synced or cached.
+  - [ET-1192] `GroupedContacts.item` was renamed to `items`
+  - [ET-1192] `ContactGroupItem.email` was renamed to `contacts`
+    and now carries `Vec<ContactItem>` instead of `Vec<ContactEmailItem>`
 
 ### Fixed